// types/error.ts
export enum ErrorMessage {
  PO_EXISTS = 'PO_EXISTS',
  BOOKING_EXISTS = 'BOOKING_EXISTS',
  USER_ACCOUNT_NOT_FOUND = 'USER_ACCOUNT_NOT_FOUND',
  INVALID_OR_EXPIRED_TOKEN = 'INVALID_OR_EXPIRED_TOKEN',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  INCORRECT_DETAILS = 'INCORRECT_DETAILS',
  UNAUTHORIZED = 'UNAUTHORIZED',
  PAYMENT_REQUIRED = 'PAYMENT_REQUIRED',
  FORBIDDEN = 'FORBIDDEN',
  NOT_FOUND = 'NOT_FOUND',
  CONFLICT = 'CONFLICT',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
  MISSING_TOKEN = 'MISSING_TOKEN',
  MISSING_PARAMS = 'MISSING_PARAMS',
  MISSING_DATA = 'MISSING_DATA',
  SHORT_LENGTH = 'SHORT_LENGTH'
  }